refactor(racingcar): 일급 컬렉션 기능을 Racing 클래스로 통합
- Cars 클래스 삭제
